[Move] #5 - about me 상수파일 Constants 폴더로 이동
